+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 1.3.0.99
 
+* `melt.data.frame()` preserves shared classes for measure variables.
+
 * `melt.data.frame()` gains an internal Rcpp / C++ implementation, and
   is now many orders of magnitudes faster. (Thanks to Kevin Ushey)